Move Finder Scope to ui-components package and continue working on fleshing that package out.
