--- research-app/src/search.ts
@@ -1,6 +1,0 @@
-import { Place } from "@wwtelescope/engine";
-
-
-export interface SearchDataProvider {
-  placeForLocation(location: { raDeg: number; decDeg: number; }): Promise<Place | null>;
-}
