+++ ui-components/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  extends: [
+    "../.eslintrc.js",
+    "plugin:vue/essential",
+    "@vue/typescript/recommended"
+  ]
+};
