+++ ui-components/package.json
@@ -11,6 +11,20 @@
     "vue": "^3"
   },
   "description": "UI components build on top of the WWT WebGL Vue/Pinia store",
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.38.1",
+    "@typescript-eslint/parser": "^5.38.1",
+    "@vue/cli-plugin-eslint": "^5.0.8",
+    "@vue/cli-plugin-typescript": "^5.0.8",
+    "@vue/cli-service": "^5.0.8",
+    "@vue/eslint-config-typescript": "^11.0.2",
+    "eslint": "^8.24.0",
+    "eslint-plugin-vue": "^9.5.1",
+    "rimraf": "^3",
+    "typedoc": "0.25.1",
+    "typescript": "~4.8.4",
+    "vue-template-compiler": "~2.6.11"
+  },
   "homepage": "https://worldwidetelescope.org/home/",
   "keywords": [
     "WorldWide Telescope"
