+++ ui-components/src/index.ts
@@ -1,0 +1,3 @@
+export { SearchDataProvider } from "./search";
+export { DefaultSearchDataProvider } from "./default_search";
+export { default as FinderScope } from "./FinderScope.vue";
