--- ui-components/src/types.ts
@@ -1,7 +1,0 @@
-import { SearchDataProvider } from "./search";
-
-export interface FinderScopeProps {
-  modelValue: boolean;
-  searchProvider: SearchDataProvider;
-  position: [number, number];
-}
