+++ ui-components/tsconfig.json
@@ -1,0 +1,47 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "module": "esnext",
+    "composite": true,
+    "strict": true,
+    "jsx": "preserve",
+    "importHelpers": true,
+    "downlevelIteration": true,
+    "moduleResolution": "node",
+    "experimentalDecorators": true,
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "sourceMap": true,
+    "outDir": "dist"
+  },
+  "include": [
+    "src/**/*.ts",
+    "src/**/*.vue"
+  ],
+  "references": [
+    {
+      "path": "../astro"
+    },
+    {
+      "path": "../engine"
+    },
+    {
+      "path": "../engine-helpers"
+    },
+    {
+      "path": "../engine-types"
+    },
+  ],
+  "typedocOptions": {
+    "readme": "../docs/engine/engine-pinia-index.md",
+    "out": "../docs/engine/static/apiref/engine-pinia",
+    "gaID": "UA-107473046-3",
+    "excludePrivate": true,
+    "plugin": [
+      "typedoc-plugin-vue"
+    ],
+    "validation": {
+      "invalidLink": true
+    }
+  }
+}
